Split domain upstream client protocols
